Remove cheaty .length() with PaleolithicNatural
As the declaration states, the constructor `PaleolithicNatural(unsigned)` is purposed to take an `unsigned` as an interface "for the Renissance person"

This PR removes all such horrible usages of the helper interface within the implementation by changing multiplication, comparison, `half()` and `is_odd()` of the aforementioned unsigned integral type. These functions now use concepts which do not involve length, as it is accomplishes nothing to speak about the length of a an equal count of what you're trying to count in the first place. We implement parity by dividing our dots into two piles etc.

You may not want to merge this into the main branch though, as people coming from your talk may want to see your original implementation, but this should be merged into any current production environment as soon as possible to avoid any terrifying mishappenings.